chore(sonarlint): remove code flaw
had to do with int promotion and byte shifting, see https://sonarcloud.io/organizations/limechain/rules?open=java%3AS3034&rule_key=java%3AS3034

+++ src/main/java/com/limechain/trie/structure/nibble/NibblesUtils.java
@@ -54,7 +54,7 @@ public class NibblesUtils {
     }
 
     private byte nibblesToByte(Nibble first, Nibble second) {
-        return (byte) (first.asByte() << 4 | second.asByte());
+        return (byte) (first.asInt() << 4 | second.asInt());
     }
 
     private enum ConversionStrategy {
